fix: make encoder output GOARCH-reproducible via float32 FMA barriers (#78)
Adds explicit float32() rounding barriers at the encoder's float32 a*b+c
sites so the Go compiler cannot contract them into a fused multiply-add on
arm64 (it does not on default amd64), matching the -ffp-contract=off FFmpeg
reference. The encoder now produces byte-identical output on arm64 and amd64
given identical input; the barriers are a no-op on amd64.

Adds two guards: a build-time scan that fails on any float32 FMA, and a
committed cross-arch ADTS determinism golden that CI runs on both arches.

The originally reported corpus splits were partly a harness artifact: the
tonal signal generator's float64 math is FMA-contracted on arm64, so its
float32 input samples differ across arches independent of the encoder.

Closes #59.

+++ internal/psy/lame.go
@@ -18,9 +18,9 @@ func psyHpFilter(firbuf []float32, hpfsmpl *[blockSizeLong]float32) {
 		sum1 = firbuf[i+(psyLameFirLen-1)/2]
 		sum2 = 0.0
 		for j := 0; j < ((psyLameFirLen-1)/2)-1; j += 2 {
-			t1 := tables.PsyFirCoeffs[j] * (firbuf[i+j] + firbuf[i+psyLameFirLen-j])
+			t1 := float32(tables.PsyFirCoeffs[j] * (firbuf[i+j] + firbuf[i+psyLameFirLen-j])) // no cross-statement FMA
 			sum1 += t1
-			t2 := tables.PsyFirCoeffs[j+1] * (firbuf[i+j+1] + firbuf[i+psyLameFirLen-j-1])
+			t2 := float32(tables.PsyFirCoeffs[j+1] * (firbuf[i+j+1] + firbuf[i+psyLameFirLen-j-1])) // no cross-statement FMA
 			sum2 += t2
 		}
 		hpfsmpl[i] = (sum1 + sum2) * 32768.0
